Support scons parallel build jobs
Contrail source issuing a cycle compilation issue when we build source
in parallel and that needs to be fixed in the scons scripts. Until it
works, use that ugly workaround which retry five times to build
contrail with jobs in parallel before abandon and fail.
Also use parallel to compile cassandra C++ drivers.